feat: Add newEnvio function to entregas API module
This commit adds a new function, newEnvio, to the entregas API module. The newEnvio function is responsible for making a POST request to the 'entregas/envio/{id}' endpoint, with the provided entrega data. This function allows for creating a new envio (delivery) associated with a specific id. The response from the API is returned.